msm: kgsl: Don't allocate memory dynamically for temp command buffers
The temporary command buffer in _set_pagetable_gpu is only the size of a
single page, and _set_pagetable_gpu is never executed concurrently. It
is therefore easy to replace the dynamic command buffer allocation with
a static one to improve performance by avoiding the latency of dynamic
memory allocation.

Change-Id: I4f053f2b1f6f9737ea89edb34da020bc7d9df96d
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>